Remove Flux.singleOptional()
