--- reactor-core/src/main/java/reactor/core/publisher/MonoSingleOptionalMono.java
@@ -1,47 +1,0 @@
-/*
- * Copyright (c) 2016-2021 VMware Inc. or its affiliates, All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package reactor.core.publisher;
-
-import reactor.core.CoreSubscriber;
-
-import java.util.Optional;
-
-/**
- * Wraps the item from the source into an Optional, emits
- * an empty Optional instead for empty source or signals
- * IndexOutOfBoundsException for a multi-item source.
- *
- * @param <T> the value type
- * @see <a href="https://github.com/reactor/reactive-streams-commons">Reactive-Streams-Commons</a>
- */
-final class MonoSingleOptionalMono<T> extends InternalMonoOperator<T, Optional<T>> {
-
-	MonoSingleOptionalMono(Mono<? extends T> source) {
-		super(source);
-	}
-
-	@Override
-	public CoreSubscriber<? super T> subscribeOrReturn(CoreSubscriber<? super Optional<T>> actual) {
-		return new MonoSingleOptional.SingleOptionalSubscriber<>(actual);
-	}
-
-	@Override
-	public Object scanUnsafe(Attr key) {
-		if (key == Attr.RUN_STYLE) return Attr.RunStyle.SYNC;
-		return super.scanUnsafe(key);
-	}
-}
