+++ reactor-core/src/test/java/reactor/core/publisher/MonoSingleOptionalTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2021 VMware Inc. or its affiliates, All Rights Reserved.
+ * Copyright (c) 2017-2021 VMware Inc. or its affiliates, All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,275 +16,163 @@
 
 package reactor.core.publisher;
 
-import java.util.NoSuchElementException;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatCode;
+import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
+
 import java.util.Optional;
 import java.util.concurrent.Callable;
+import java.util.function.Function;
 
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.reactivestreams.Subscription;
+
 import reactor.core.CoreSubscriber;
 import reactor.core.Fuseable;
 import reactor.core.Scannable;
 import reactor.test.StepVerifier;
-import reactor.test.subscriber.AssertSubscriber;
-
-import static org.assertj.core.api.Assertions.*;
 
 public class MonoSingleOptionalTest {
 
-	@Nested
-	class ConcreteClassConsistency {
-		//tests Flux.singleOptional, and Mono.singleOptional API consistency over returned classes
+    @Nested
+    class ConcreteClassConsistency {
+        // tests Mono.singleOptional returned classes
 
-		@Test
-		void monoWithScalarEmpty() {
-			Mono<Integer> source = Mono.empty();
-			Mono<Optional<Integer>> singleOptional = source.singleOptional();
+        @Test
+        void monoWithScalarEmpty() {
+            Mono<Integer> source = Mono.empty();
+            Mono<Optional<Integer>> singleOptional = source.singleOptional();
 
-			assertThat(source).as("source").isInstanceOf(Fuseable.ScalarCallable.class);
-			assertThat(singleOptional).as("singleOptional")
-			                  .isInstanceOf(MonoJust.class)
-			                  .isInstanceOf(Fuseable.ScalarCallable.class);
-		}
+            assertThat(source).as("source").isInstanceOf(Fuseable.ScalarCallable.class);
+            assertThat(singleOptional).as("singleOptional")
+                              .isInstanceOf(MonoJust.class)
+                              .isInstanceOf(Fuseable.ScalarCallable.class);
+        }
 
-		@Test
-		void monoWithScalarError() {
-			Mono<Integer> source = Mono.error(new IllegalStateException("test"));
-			Mono<Optional<Integer>> singleOptional = source.singleOptional();
+        @Test
+        void monoWithScalarError() {
+            Mono<Integer> source = Mono.error(new IllegalStateException("test"));
+            Mono<Optional<Integer>> singleOptional = source.singleOptional();
 
-			assertThat(source).as("source").isInstanceOf(Fuseable.ScalarCallable.class);
-			assertThat(singleOptional).as("singleOptional")
-			                  .isInstanceOf(MonoError.class)
-			                  .isInstanceOf(Fuseable.ScalarCallable.class);
-		}
+            assertThat(source).as("source").isInstanceOf(Fuseable.ScalarCallable.class);
+            assertThat(singleOptional).as("singleOptional")
+                              .isInstanceOf(MonoError.class)
+                              .isInstanceOf(Fuseable.ScalarCallable.class);
+        }
 
-		@Test
-		void monoWithScalarValue() {
-			Mono<Integer> source = Mono.just(1);
-			Mono<Optional<Integer>> single = source.singleOptional();
+        @Test
+        void monoWithScalarValue() {
+            Mono<Integer> source = Mono.just(1);
+            Mono<Optional<Integer>> single = source.singleOptional();
 
-			assertThat(source).as("source").isInstanceOf(Fuseable.ScalarCallable.class);
-			assertThat(single).as("single")
-			                  .isInstanceOf(MonoJust.class)
-			                  .isInstanceOf(Fuseable.ScalarCallable.class);
-		}
+            assertThat(source).as("source").isInstanceOf(Fuseable.ScalarCallable.class);
+            assertThat(single).as("singleOptional")
+                              .isInstanceOf(MonoJust.class)
+                              .isInstanceOf(Fuseable.ScalarCallable.class);
+        }
 
-		@Test
-		void monoWithCallable() {
-			Mono<Integer> source = Mono.fromSupplier(() -> 1);
-			Mono<Optional<Integer>> single = source.singleOptional();
+        @Test
+        void monoWithCallable() {
+            Mono<Integer> source = Mono.fromSupplier(() -> 1);
+            Mono<Optional<Integer>> single = source.singleOptional();
 
-			assertThat(source).as("source")
-			                  .isInstanceOf(Callable.class)
-			                  .isNotInstanceOf(Fuseable.ScalarCallable.class);
-			assertThat(single).as("single").isInstanceOf(MonoSingleOptionalCallable.class);
-		}
+            assertThat(source).as("source")
+                              .isInstanceOf(Callable.class)
+                              .isNotInstanceOf(Fuseable.ScalarCallable.class);
+            assertThat(single).as("singleOptional").isInstanceOf(MonoSingleOptionalCallable.class);
+        }
 
-		@Test
-		void monoWithNormal() {
-			Mono<Integer> source = Mono.just(1).hide();
-			Mono<Optional<Integer>> single = source.singleOptional();
+        @Test
+        void monoWithNormal() {
+            Mono<Integer> source = Mono.just(1).hide();
+            Mono<Optional<Integer>> single = source.singleOptional();
 
-			assertThat(source).as("source").isNotInstanceOf(Callable.class); //excludes ScalarCallable too
-			assertThat(single).as("single").isInstanceOf(MonoSingleOptionalMono.class);
-		}
+            assertThat(source).as("source").isNotInstanceOf(Callable.class); //excludes ScalarCallable too
+            assertThat(single).as("singleOptional").isInstanceOf(MonoSingleOptional.class);
+        }
+    }
 
-		@Test
-		void fluxWithScalarEmpty() {
-			Flux<Integer> source = Flux.empty();
-			Mono<Optional<Integer>> single = source.singleOptional();
-
-			assertThat(source).as("source").isInstanceOf(Fuseable.ScalarCallable.class);
-			assertThat(single).as("single")
-			                  .isInstanceOf(MonoJust.class)
-			                  .isInstanceOf(Fuseable.ScalarCallable.class);
-		}
-
-		@Test
-		void fluxWithScalarError() {
-			Flux<Integer> source = Flux.error(new IllegalStateException("test"));
-			Mono<Optional<Integer>> single = source.singleOptional();
-
-			assertThat(source).as("source").isInstanceOf(Fuseable.ScalarCallable.class);
-			assertThat(single).as("single")
-			                  .isInstanceOf(MonoError.class)
-			                  .isInstanceOf(Fuseable.ScalarCallable.class);
-		}
-
-		@Test
-		void fluxWithScalarValue() {
-			Flux<Integer> source = Flux.just(1);
-			Mono<Optional<Integer>> single = source.singleOptional();
-
-			assertThat(source).as("source").isInstanceOf(Fuseable.ScalarCallable.class);
-			assertThat(single).as("single")
-			                  .isInstanceOf(MonoJust.class)
-			                  .isInstanceOf(Fuseable.ScalarCallable.class);
-		}
-
-		@Test
-		void fluxWithCallable() {
-			Flux<Integer> source = Mono.fromSupplier(() -> 1).flux();
-			Mono<Optional<Integer>> single = source.singleOptional();
-
-			assertThat(source).as("source")
-			                  .isInstanceOf(Callable.class)
-			                  .isNotInstanceOf(Fuseable.ScalarCallable.class);
-			assertThat(single).as("single").isInstanceOf(MonoSingleOptionalCallable.class);
-		}
-
-		@Test
-		void fluxWithNormal() {
-			Flux<Integer> source = Flux.range(1, 10);
-			Mono<Optional<Integer>> single = source.singleOptional();
-
-			assertThat(source).as("source").isNotInstanceOf(Callable.class); //excludes ScalarCallable too
-			assertThat(single).as("single").isInstanceOf(MonoSingleOptional.class);
-		}
-
-	}
-
-	@Test
-	void source1Null() {
-		assertThatExceptionOfType(NullPointerException.class).isThrownBy(() -> {
-			new MonoSingleOptional<>(null);
-		});
-	}
-
-	@Test
-	public void normal() {
-
-		AssertSubscriber<Optional<Integer>> ts = AssertSubscriber.create();
-
-		Flux.just(1).singleOptional().subscribe(ts);
-
-		ts.assertValues(Optional.of(1))
-		  .assertNoError()
-		  .assertComplete();
-	}
-
-	@Test
-	public void normalBackpressured() {
-		AssertSubscriber<Optional<Integer>> ts = AssertSubscriber.create(0);
-
-		Flux.just(1).singleOptional().subscribe(ts);
-
-		ts.assertNoValues()
-		  .assertNoError()
-		  .assertNotComplete();
-
-		ts.request(1);
-
-		ts.assertValues(Optional.of(1))
-		  .assertNoError()
-		  .assertComplete();
-	}
-
-	@Test
-	public void empty() {
-
-		AssertSubscriber<Optional<Integer>> ts = AssertSubscriber.create();
-
-		Flux.<Integer>empty().singleOptional().subscribe(ts);
-
-		ts.assertValues(Optional.empty())
-		  .assertNoError()
-		  .assertComplete();
-	}
-
-	@Test
-	public void error() {
-		StepVerifier.create(Flux.error(new RuntimeException("forced failure"))
-		                        .singleOptional())
-		            .verifyErrorMessage("forced failure");
-	}
-
-	@Test
-	public void errorHide() {
-		StepVerifier.create(Flux.error(new RuntimeException("forced failure"))
-		                        .hide()
-		                        .singleOptional())
-		            .verifyErrorMessage("forced failure");
-	}
-
-	@Test
-	public void multi() {
-
-		AssertSubscriber<Optional<Integer>> ts = AssertSubscriber.create();
-
-		Flux.range(1, 10).singleOptional().subscribe(ts);
-
-		ts.assertNoValues()
-		  .assertError(IndexOutOfBoundsException.class)
-		  .assertNotComplete();
-	}
-
-	@Test
-	public void multiBackpressured() {
-
-		AssertSubscriber<Optional<Integer>> ts = AssertSubscriber.create(0);
-
-		Flux.range(1, 10).singleOptional().subscribe(ts);
-
-		ts.assertNoValues()
-		  .assertNoError()
-		  .assertNotComplete();
-
-		ts.request(1);
-
-		ts.assertNoValues()
-		  .assertError(IndexOutOfBoundsException.class)
-		  .assertNotComplete();
-	}
-
-	@Test
-	public void singleCallable() {
-		StepVerifier.create(Mono.fromCallable(() -> 1)
-		                        .flux()
-		                        .singleOptional())
-		            .expectNext(Optional.of(1))
-		            .verifyComplete();
-	}
-
-	@Test
-	public void singleJustHide() {
-		StepVerifier.create(Flux.empty()
-								.hide()
-		                        .singleOptional())
+    @Test
+    void source1Null() {
+        assertThatExceptionOfType(NullPointerException.class).isThrownBy(() -> {
+            new MonoSingleOptional<>(null);
+        });
+    }
+    
+    @Test
+	public void callableEmpty() {
+		StepVerifier.create(Mono.empty().singleOptional())
 					.expectNext(Optional.empty())
+					.verifyComplete();
+	}
+
+	@Test
+	public void callableValued() {
+		StepVerifier.create(Mono.just("foo").singleOptional())
+		            .expectNext(Optional.of("foo"))
 		            .verifyComplete();
+	}
+
+	@Test
+	public void normalEmpty() {
+		StepVerifier.create(Mono.empty().hide().singleOptional())
+							.expectNext(Optional.empty())
+							.verifyComplete();
+	}
+
+	@Test
+	public void normalValued() {
+		StepVerifier.create(Mono.just("foo").hide().singleOptional())
+		            .expectNext(Optional.of("foo"))
+		            .verifyComplete();
+	}
+
+	@Test
+	void fusionMonoSingleFusion() {
+		Mono<Optional<Integer>> fusedCase = Mono
+				.just(1)
+				.map(Function.identity())
+				.singleOptional();
+
+		assertThat(fusedCase)
+				.as("fusedCase assembly check")
+				.isInstanceOf(MonoSingleOptional.class)
+				.isNotInstanceOf(Fuseable.class);
+
+		assertThatCode(() -> fusedCase.filter(v -> true).block())
+				.as("fusedCase fused")
+				.doesNotThrowAnyException();
 	}
 
 	@Test
 	public void scanOperator(){
-	    MonoSingleOptional<String> test = new MonoSingleOptional<>(Flux.just("foo"));
+	    MonoSingleOptional<String> test = new MonoSingleOptional<>(Mono.just("foo"));
 
 	    assertThat(test.scan(Scannable.Attr.RUN_STYLE)).isSameAs(Scannable.Attr.RunStyle.SYNC);
 	}
 
-	@Test
-	public void scanSubscriber() {
-		CoreSubscriber<Optional<String>>
-				actual = new LambdaMonoSubscriber<>(null, e -> {}, null, null);
-		MonoSingleOptional.SingleOptionalSubscriber<String> test = new MonoSingleOptional.SingleOptionalSubscriber<>(
-				actual);
-		Subscription parent = Operators.emptySubscription();
-		test.onSubscribe(parent);
+    @Test
+    public void scanSubscriber() {
+        CoreSubscriber<Optional<String>>
+                actual = new LambdaMonoSubscriber<>(null, e -> {}, null, null);
+        MonoSingleOptional.SingleOptionalSubscriber<String> test = new MonoSingleOptional.SingleOptionalSubscriber<>(
+                actual);
+        Subscription parent = Operators.emptySubscription();
+        test.onSubscribe(parent);
 
-		assertThat(test.scan(Scannable.Attr.PREFETCH)).isEqualTo(Integer.MAX_VALUE);
+        assertThat(test.scan(Scannable.Attr.PREFETCH)).isEqualTo(Integer.MAX_VALUE);
 
-		assertThat(test.scan(Scannable.Attr.PARENT)).isSameAs(parent);
-		assertThat(test.scan(Scannable.Attr.ACTUAL)).isSameAs(actual);
-		assertThat(test.scan(Scannable.Attr.RUN_STYLE)).isSameAs(Scannable.Attr.RunStyle.SYNC);
+        assertThat(test.scan(Scannable.Attr.PARENT)).isSameAs(parent);
+        assertThat(test.scan(Scannable.Attr.ACTUAL)).isSameAs(actual);
+        assertThat(test.scan(Scannable.Attr.RUN_STYLE)).isSameAs(Scannable.Attr.RunStyle.SYNC);
 
-		assertThat(test.scan(Scannable.Attr.TERMINATED)).isFalse();
-		test.onError(new IllegalStateException("boom"));
-		assertThat(test.scan(Scannable.Attr.TERMINATED)).isTrue();
+        assertThat(test.scan(Scannable.Attr.TERMINATED)).isFalse();
+        test.onError(new IllegalStateException("boom"));
+        assertThat(test.scan(Scannable.Attr.TERMINATED)).isTrue();
 
-		assertThat(test.scan(Scannable.Attr.CANCELLED)).isFalse();
-		test.cancel();
-		assertThat(test.scan(Scannable.Attr.CANCELLED)).isTrue();
-	}
-
+        assertThat(test.scan(Scannable.Attr.CANCELLED)).isFalse();
+        test.cancel();
+        assertThat(test.scan(Scannable.Attr.CANCELLED)).isTrue();
+    }
 }
